requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.2
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
h11==0.14.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.2.2
pytest-django==4.8.0
sqlparse==0.5.0
tomli==2.0.1
tzdata==2024.1
Werkzeug==3.0.3
setuptools==70.2.0
cryptography==42.0.8
Python==3.12.4